Shrunk move struct score field to 16 bits.

+++ src/heuristics.rs
@@ -1,13 +1,13 @@
 use crate::piece::Piece;
 
-const MVV_LVA_TAB: [[i32; 6]; 6] = init_mvv_lva_table();
+const MVV_LVA_TAB: [[i16; 6]; 6] = init_mvv_lva_table();
 
-const fn init_mvv_lva_table() -> [[i32; 6]; 6] {
+const fn init_mvv_lva_table() -> [[i16; 6]; 6] {
     let mut tab = [[0; 6]; 6];
     let mut a = 0; let mut v = 0;
     while a < 6 {
         while v < 6 {
-            tab[v][a] = (((v + 1) * 10) - (6 - a)) as i32;
+            tab[v][a] = (((v + 1) * 10) - (6 - a)) as i16;
             v += 1;
         }
         a += 1;
@@ -15,6 +15,6 @@ const fn init_mvv_lva_table() -> [[i32; 6]; 6] {
     tab
 }
 
-pub fn calc_mvv_lva_heuristic(piece: Piece, enemy_piece: Piece) -> i32 {
+pub fn calc_mvv_lva_heuristic(piece: Piece, enemy_piece: Piece) -> i16 {
     MVV_LVA_TAB[enemy_piece as usize][piece as usize]
 }
